+++ src/crawlee/http_clients/_httpx.py
@@ -6,7 +6,7 @@ import httpx
 from typing_extensions import override
 
 from crawlee._utils.blocked import ROTATE_PROXY_ERRORS
-from crawlee.errors import HttpStatusCodeError, ProxyError
+from crawlee.errors import ProxyError
 from crawlee.http_clients import BaseHttpClient, HttpCrawlingResult, HttpResponse
 from crawlee.sessions import Session
 
@@ -66,6 +66,8 @@ class HttpxHttpClient(BaseHttpClient):
 
     This client uses the `HTTPX` library to perform HTTP requests in crawlers (`BasicCrawler` subclasses)
     and to manage sessions, proxies, and error handling.
+
+    See the `BaseHttpClient` class for more common information about HTTP clients.
     """
 
     def __init__(
@@ -84,9 +86,11 @@ class HttpxHttpClient(BaseHttpClient):
             ignore_http_error_status_codes: HTTP status codes to ignore as errors.
             async_client_kwargs: Additional keyword arguments for `httpx.AsyncClient`.
         """
-        self._persist_cookies_per_session = persist_cookies_per_session
-        self._additional_http_error_status_codes = set(additional_http_error_status_codes)
-        self._ignore_http_error_status_codes = set(ignore_http_error_status_codes)
+        super().__init__(
+            persist_cookies_per_session=persist_cookies_per_session,
+            additional_http_error_status_codes=additional_http_error_status_codes,
+            ignore_http_error_status_codes=ignore_http_error_status_codes,
+        )
         self._async_client_kwargs = async_client_kwargs
 
         self._client_by_proxy_url = dict[Optional[str], httpx.AsyncClient]()
@@ -122,16 +126,11 @@ class HttpxHttpClient(BaseHttpClient):
         if statistics:
             statistics.register_status_code(response.status_code)
 
-        exclude_error = response.status_code in self._ignore_http_error_status_codes
-        include_error = response.status_code in self._additional_http_error_status_codes
-
-        if include_error or (self._is_server_code(response.status_code) and not exclude_error):
-            if include_error:
-                raise HttpStatusCodeError(
-                    f'Status code {response.status_code} (user-configured to be an error) returned'
-                )
-
-            raise HttpStatusCodeError(f'Status code {response.status_code} returned')
+        self._raise_for_error_status_code(
+            response.status_code,
+            self._additional_http_error_status_codes,
+            self._ignore_http_error_status_codes,
+        )
 
         request.loaded_url = str(response.url)
 
@@ -168,6 +167,12 @@ class HttpxHttpClient(BaseHttpClient):
             if self._is_proxy_error(exc):
                 raise ProxyError from exc
             raise
+
+        self._raise_for_error_status_code(
+            response.status_code,
+            self._additional_http_error_status_codes,
+            self._ignore_http_error_status_codes,
+        )
 
         return _HttpxResponse(response)