+++ tests/unit/_autoscaling/test_system_status.py
@@ -20,7 +20,9 @@ from crawlee.events import LocalEventManager
 
 @pytest.fixture
 async def snapshotter() -> AsyncGenerator[Snapshotter, None]:
-    async with LocalEventManager() as event_manager, Snapshotter(event_manager) as snapshotter:
+    async with LocalEventManager() as event_manager, Snapshotter(
+        event_manager, available_memory_ratio=0.25
+    ) as snapshotter:
         yield snapshotter
 
 
@@ -30,7 +32,9 @@ def now() -> datetime:
 
 
 async def test_start_stop_lifecycle() -> None:
-    async with LocalEventManager() as event_manager, Snapshotter(event_manager) as snapshotter:
+    async with LocalEventManager() as event_manager, Snapshotter(
+        event_manager, available_memory_ratio=0.25
+    ) as snapshotter:
         system_status = SystemStatus(snapshotter)
         system_status.get_current_system_info()
         system_status.get_historical_system_info()